Move bin.ts core into -cli/run.ts
